fix some checks

+++ code/Celestial.cpp
@@ -24,7 +24,7 @@ void Celestial::computeGravity(SpaceShip& player) {
     player.computeGravity(physics.getPosition(),physics.getMass(),2000);
 }
 void Celestial::CheckHit(std::vector<Bullet>& bullet,std::vector<Celestial*>& celestials,int ind) {
-    for (int i=0;i<bullet.size();i++) {
+    for (long unsigned int i=0;i<bullet.size();i++) {
         if (collider.isCollidingWith(physics,bullet[i].getPhysics(),bullet[i].getCollider())) {
             LoseHealth(bullet[i].getDamage(),celestials,ind);
             bullet.erase(bullet.begin()+i);
