+++ code/SpaceShip.cpp
@@ -87,7 +87,7 @@ void SpaceShip::alignToPlanet(const Physics& planetPhys) {
 
 }
 void SpaceShip::UpdateBullets(sf::Time dt,sf::RenderWindow& window,sf::FloatRect& viewRect) {
-    for (int i=0;i<bullets.size();i++) {
+    for (long unsigned int i=0;i<bullets.size();i++) {
         bullets[i].Update(dt,getPhysics().getPosition(),window,viewRect);
         if (bullets[i].getLifetime()<0) {
             bullets.erase(bullets.begin()+i);
